# gatewayDemo/.history/controller/service_20201119222041.go
package controller

import "github.com/gin-gonic/gin"

type ServiceController struct {
}

func ServiceRegister(group *gin.RouterGroup) {
	service := &ServiceController{}
	group.GET("/service_list")
}

// ServiceList godoc
// @Summary 服务列表
// @Description 服务列表
// @Tags 服务管理
// @ID /service/service_list
// @Accept  json
// @Produce  json
// @Param body body string true "关键词"
// @Param body body string true "页数"
// @Param body body string true "每页条数"
// @Success 200 {object} middleware.Response{data=dto.AdminLoginInput} "success"
// @Router /service/service_list [get]
func (service *ServiceController) ServiceList(c *gin.Context) {
	servicelist := 
}